refactor: empty feature add strategy duplication
